Fix DTrace-enabled builds on gcc 4.5

+++ src/share/vm/prims/jni.cpp
@@ -2804,7 +2804,7 @@ extern "C" jobject JNICALL jni_NewDirectByteBuffer(JNIEnv *env, void* address, j
 
   if (!directBufferSupportInitializeEnded) {
     if (!initializeDirectBufferSupport(env, thread)) {
-      DTRACE_PROBE1(hotspot_jni, NewDirectByteBuffer__return, NULL);
+      DTRACE_PROBE1(hotspot_jni, NewDirectByteBuffer__return, (uintptr_t) NULL);
       return NULL;
     }
   }
